Fix process return code in case of error
Twitter stream client errors are handler through process restart, using a supervisor.
Since systemd migration, the systemd unit handled auto restarts only if process exited with an error (returnCode != 0), this was not the case and thus the service was definitely stopped.
This commit fixes this by returning a non 0 status code if the twitter stream client fails.